+++ packages/plugins/history-manager/extension.ts
@@ -1,16 +1,27 @@
 import SQLTools from '@sqltools/core/plugin-api';
-import { ExtensionContext } from 'vscode';
 import HistoryExplorer from './explorer';
+import { DatabaseInterface } from '@sqltools/core/interface';
+import { getNameFromId } from '@sqltools/core/utils';
+
+const hookedCommands = [
+  'executeFromInput',
+  'executeQuery',
+  'executeQueryFromFile',
+];
 
 export default class ConnectionManagerPlugin implements SQLTools.ExtensionPlugin {
-  private context: ExtensionContext;
   private explorer: HistoryExplorer;
+  private addToHistoryHook = (evt: SQLTools.CommandSuccessEvent<DatabaseInterface.QueryResults[]>) => {
+    evt.result.forEach(r => {
+      this.explorer.addItem(getNameFromId(r.connId), r.query);
+    });
+  }
 
   public register(extension: SQLTools.ExtensionInterface) {
-    if (this.context) return; // do not register twice
-    this.context = extension.context;
+    if (this.explorer) return; // do not register twice
 
-    this.explorer = new HistoryExplorer();
-    this.explorer.refresh();
+    this.explorer = new HistoryExplorer(extension.context);
+
+    hookedCommands.forEach(cmd => extension.addAfterCommandSuccessHook(cmd, this.addToHistoryHook))
   }
 }